feat(ui): enhance theme system and reorganize project structure
- Reorganize HomeScreen.kt to proper directory structure
- Update NavGraph.kt imports for new screen locations
- Adjust splashTitle fontSize from 48sp to 60sp for better visibility
- Fine-tune SplashBgGradient colors for improved visual appeal
- Enhance theme consistency across splash screen components
